test(api): transform add-oidc-provider usecase unit tests to integration tests
